v0.0.1a3/mybot/settings.py
from decouple import config

HOSTNAME = ''

PLATFORMS = {
    'telegram': {
        'ENGINE': 'bottery.platform.telegram',
        'OPTIONS': {
            'token': config('TOKEN'),
        }
    },
}
