v0.0.1a3/requirements.txt
bottery==0.0.1a3
python-decouple
